Disallow deleting others' annotations by default.
This introduces SessionConstraints into SessionController
and AnnotationController. The latter comes with a new
permission, namely "edit.annotations.allowNonAuthorDelete".

Code is added into AnnotationController to check
for the author name of the annotation node to be deleted,
and do nothing if it matches the name of the current
user.

Inside EditorSession which instantiates the SessionConstraints
instance, "edit.annotations.allowNonAuthorDelete" is disabled
to achieve the desired result.